requirements.txt
joblib~=1.0.1
numpy~=1.20.1
farm-haystack~=0.7.0